+++ plugins/source/aws/client/helpers.go
@@ -7,7 +7,6 @@ import (
 	"errors"
 	"fmt"
 	"reflect"
-	"regexp"
 	"strings"
 	"sync"
 	"time"
@@ -93,7 +92,7 @@ var accessDeniedErrorStrings = map[string]struct{}{
 	"Unauthorized":                    {},
 }
 
-func readSupportedServiceRegions() *SupportedServiceRegionsData {
+func ReadSupportedServiceRegions() *SupportedServiceRegionsData {
 	f, err := supportedServiceRegionFile.Open(PartitionServiceRegionFile)
 	if err != nil {
 		return nil
@@ -126,7 +125,7 @@ func readSupportedServiceRegions() *SupportedServiceRegionsData {
 
 func isSupportedServiceForRegion(service string, region string) bool {
 	readOnce.Do(func() {
-		supportedServiceRegion = readSupportedServiceRegions()
+		supportedServiceRegion = ReadSupportedServiceRegions()
 	})
 
 	if supportedServiceRegion == nil {
@@ -153,7 +152,7 @@ func isSupportedServiceForRegion(service string, region string) bool {
 
 func getAvailableRegions() (map[string]bool, error) {
 	readOnce.Do(func() {
-		supportedServiceRegion = readSupportedServiceRegions()
+		supportedServiceRegion = ReadSupportedServiceRegions()
 	})
 
 	regionsSet := make(map[string]bool)
@@ -179,7 +178,7 @@ func getAvailableRegions() (map[string]bool, error) {
 
 func RegionsPartition(region string) (string, bool) {
 	readOnce.Do(func() {
-		supportedServiceRegion = readSupportedServiceRegions()
+		supportedServiceRegion = ReadSupportedServiceRegions()
 	})
 
 	prt, ok := supportedServiceRegion.regionVsPartition[region]
@@ -231,19 +230,6 @@ func IgnoreNotAvailableRegion(err error) bool {
 	return false
 }
 
-// makeARN creates an ARN using supplied service name, partition, account id, region name and resource id parts.
-// Resource id parts are concatenated using forward slash (/).
-// See https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html for more information.
-func makeARN(service AWSService, partition, accountID, region string, idParts ...string) arn.ARN {
-	return arn.ARN{
-		Partition: partition,
-		Service:   string(service),
-		Region:    region,
-		AccountID: accountID,
-		Resource:  strings.Join(idParts, "/"),
-	}
-}
-
 func resolveARN(service AWSService, resourceID func(resource *schema.Resource) ([]string, error), useRegion, useAccountID bool) schema.ColumnResolver {
 	return func(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
 		cl := meta.(*Client)
@@ -258,7 +244,13 @@ func resolveARN(service AWSService, resourceID func(resource *schema.Resource) (
 		if useRegion {
 			region = cl.Region
 		}
-		return resource.Set(c.Name, makeARN(service, cl.Partition, accountID, region, idParts...).String())
+		return resource.Set(c.Name, arn.ARN{
+			Partition: cl.Partition,
+			Service:   string(service),
+			Region:    region,
+			AccountID: accountID,
+			Resource:  strings.Join(idParts, "/"),
+		}.String())
 	}
 }
 
@@ -345,17 +337,6 @@ func IsAWSError(err error, code ...string) bool {
 		if strings.Contains(ae.ErrorCode(), c) {
 			return true
 		}
-	}
-	return false
-}
-
-func IsErrorRegex(err error, code string, messageRegex *regexp.Regexp) bool {
-	var ae smithy.APIError
-	if !errors.As(err, &ae) {
-		return false
-	}
-	if ae.ErrorCode() == code && messageRegex.MatchString(ae.ErrorMessage()) {
-		return true
 	}
 	return false
 }