+++ plugins/source/aws/client/services.go
@@ -4,6 +4,7 @@ package client
 import (
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/accessanalyzer"
+	"github.com/aws/aws-sdk-go-v2/service/account"
 	"github.com/aws/aws-sdk-go-v2/service/acm"
 	"github.com/aws/aws-sdk-go-v2/service/apigateway"
 	"github.com/aws/aws-sdk-go-v2/service/apigatewayv2"
@@ -54,6 +55,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/inspector"
 	"github.com/aws/aws-sdk-go-v2/service/inspector2"
 	"github.com/aws/aws-sdk-go-v2/service/iot"
+	"github.com/aws/aws-sdk-go-v2/service/kafka"
 	"github.com/aws/aws-sdk-go-v2/service/kinesis"
 	"github.com/aws/aws-sdk-go-v2/service/kms"
 	"github.com/aws/aws-sdk-go-v2/service/lambda"
@@ -73,11 +75,13 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/aws/aws-sdk-go-v2/service/s3control"
 	"github.com/aws/aws-sdk-go-v2/service/sagemaker"
+	"github.com/aws/aws-sdk-go-v2/service/scheduler"
 	"github.com/aws/aws-sdk-go-v2/service/secretsmanager"
 	"github.com/aws/aws-sdk-go-v2/service/servicecatalog"
 	"github.com/aws/aws-sdk-go-v2/service/servicecatalogappregistry"
 	"github.com/aws/aws-sdk-go-v2/service/servicequotas"
 	"github.com/aws/aws-sdk-go-v2/service/sesv2"
+	"github.com/aws/aws-sdk-go-v2/service/sfn"
 	"github.com/aws/aws-sdk-go-v2/service/shield"
 	"github.com/aws/aws-sdk-go-v2/service/sns"
 	"github.com/aws/aws-sdk-go-v2/service/sqs"
@@ -99,6 +103,7 @@ func initServices(region string, c aws.Config) Services {
 	awsCfg.Region = region
 	return Services{
 		Accessanalyzer:            accessanalyzer.NewFromConfig(awsCfg),
+		Account:                   account.NewFromConfig(awsCfg),
 		Acm:                       acm.NewFromConfig(awsCfg),
 		Apigateway:                apigateway.NewFromConfig(awsCfg),
 		Apigatewayv2:              apigatewayv2.NewFromConfig(awsCfg),
@@ -149,6 +154,7 @@ func initServices(region string, c aws.Config) Services {
 		Inspector:                 inspector.NewFromConfig(awsCfg),
 		Inspector2:                inspector2.NewFromConfig(awsCfg),
 		Iot:                       iot.NewFromConfig(awsCfg),
+		Kafka:                     kafka.NewFromConfig(awsCfg),
 		Kinesis:                   kinesis.NewFromConfig(awsCfg),
 		Kms:                       kms.NewFromConfig(awsCfg),
 		Lambda:                    lambda.NewFromConfig(awsCfg),
@@ -168,11 +174,13 @@ func initServices(region string, c aws.Config) Services {
 		S3:                        s3.NewFromConfig(awsCfg),
 		S3control:                 s3control.NewFromConfig(awsCfg),
 		Sagemaker:                 sagemaker.NewFromConfig(awsCfg),
+		Scheduler:                 scheduler.NewFromConfig(awsCfg),
 		Secretsmanager:            secretsmanager.NewFromConfig(awsCfg),
 		Servicecatalog:            servicecatalog.NewFromConfig(awsCfg),
 		Servicecatalogappregistry: servicecatalogappregistry.NewFromConfig(awsCfg),
 		Servicequotas:             servicequotas.NewFromConfig(awsCfg),
 		Sesv2:                     sesv2.NewFromConfig(awsCfg),
+		Sfn:                       sfn.NewFromConfig(awsCfg),
 		Shield:                    shield.NewFromConfig(awsCfg),
 		Sns:                       sns.NewFromConfig(awsCfg),
 		Sqs:                       sqs.NewFromConfig(awsCfg),
@@ -191,6 +199,7 @@ func initServices(region string, c aws.Config) Services {
 
 type Services struct {
 	Accessanalyzer            services.AccessanalyzerClient
+	Account                   services.AccountClient
 	Acm                       services.AcmClient
 	Apigateway                services.ApigatewayClient
 	Apigatewayv2              services.Apigatewayv2Client
@@ -241,6 +250,7 @@ type Services struct {
 	Inspector                 services.InspectorClient
 	Inspector2                services.Inspector2Client
 	Iot                       services.IotClient
+	Kafka                     services.KafkaClient
 	Kinesis                   services.KinesisClient
 	Kms                       services.KmsClient
 	Lambda                    services.LambdaClient
@@ -260,11 +270,13 @@ type Services struct {
 	S3                        services.S3Client
 	S3control                 services.S3controlClient
 	Sagemaker                 services.SagemakerClient
+	Scheduler                 services.SchedulerClient
 	Secretsmanager            services.SecretsmanagerClient
 	Servicecatalog            services.ServicecatalogClient
 	Servicecatalogappregistry services.ServicecatalogappregistryClient
 	Servicequotas             services.ServicequotasClient
 	Sesv2                     services.Sesv2Client
+	Sfn                       services.SfnClient
 	Shield                    services.ShieldClient
 	Sns                       services.SnsClient
 	Sqs                       services.SqsClient