+++ plugins/source/aws/codegen/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"fmt"
 	"log"
 
 	"github.com/cloudquery/cloudquery/plugins/source/aws/codegen/recipes"
@@ -11,6 +12,7 @@ import (
 func generateResources() ([]*recipes.Resource, error) {
 	var resources []*recipes.Resource
 	resources = append(resources, recipes.AccessAnalyzerResources()...)
+	resources = append(resources, recipes.AccountResources()...)
 	resources = append(resources, recipes.ACMResources()...)
 	resources = append(resources, recipes.APIGatewayResources()...)
 	resources = append(resources, recipes.APIGatewayV2Resources()...)
@@ -60,6 +62,7 @@ func generateResources() ([]*recipes.Resource, error) {
 	resources = append(resources, recipes.Inspector2Resources()...)
 	resources = append(resources, recipes.InspectorResources()...)
 	resources = append(resources, recipes.IOTResources()...)
+	resources = append(resources, recipes.KafkaResources()...)
 	resources = append(resources, recipes.KinesisResources()...)
 	resources = append(resources, recipes.KMSResources()...)
 	resources = append(resources, recipes.LambdaResources()...)
@@ -77,6 +80,7 @@ func generateResources() ([]*recipes.Resource, error) {
 	resources = append(resources, recipes.Route53Resources()...)
 	resources = append(resources, recipes.S3Resources()...)
 	resources = append(resources, recipes.SagemakerResources()...)
+	resources = append(resources, recipes.SchedulerResources()...)
 	resources = append(resources, recipes.SecretsManagerResources()...)
 	resources = append(resources, recipes.ServiceCatalogResources()...)
 	resources = append(resources, recipes.ServiceQuotasResources()...)
@@ -86,6 +90,7 @@ func generateResources() ([]*recipes.Resource, error) {
 	resources = append(resources, recipes.SQSResources()...)
 	resources = append(resources, recipes.SSMResources()...)
 	resources = append(resources, recipes.SSOAdminResources()...)
+	resources = append(resources, recipes.StepFunctionResources()...)
 	resources = append(resources, recipes.TimestreamResources()...)
 	resources = append(resources, recipes.TransferResources()...)
 	resources = append(resources, recipes.WAFRegionalResources()...)
@@ -93,6 +98,11 @@ func generateResources() ([]*recipes.Resource, error) {
 	resources = append(resources, recipes.WAFv2Resources()...)
 	resources = append(resources, recipes.WorkspacesResources()...)
 	resources = append(resources, recipes.XRayResources()...)
+
+	err := recipes.SetParentChildRelationships(resources)
+	if err != nil {
+		return nil, fmt.Errorf("failed to set parent-child relationships: %w", err)
+	}
 	for _, resource := range resources {
 		if err := resource.Generate(); err != nil {
 			return nil, err