+++ plugins/source/aws/codegen/recipes/base.go
@@ -3,16 +3,20 @@ package recipes
 import (
 	"bytes"
 	"embed"
+	"errors"
 	"fmt"
 	"go/format"
 	"os"
 	"path"
 	"reflect"
+	"regexp"
 	"runtime"
 	"strings"
 	"text/template"
 
 	"github.com/aws/aws-sdk-go-v2/service/autoscaling/types"
+	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
+	"github.com/cloudquery/cloudquery/plugins/source/aws/codegen/recipes/discover"
 	"github.com/cloudquery/plugin-sdk/caser"
 	"github.com/cloudquery/plugin-sdk/codegen"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -41,11 +45,21 @@ type Resource struct {
 
 	// Used for generating the resolver and mock tests.
 	// --------------------------------
-	ShouldGenerateResolverAndMockTest bool
-	ResolverAndMockTestTemplate       string
+	ShouldGenerateResolverAndMockTest bool        // if true, resolver and mock will be generated using the options below
+	ResolverAndMockTestTemplate       string      // required: name of template directory to use
+	Client                            interface{} // required: AWS client struct to use, e.g. &ec2.Client{}
 
-	// To be used in list/paginator resolver
-	MaxResults int
+	// Applies only to list resources:
+	ListMethodName  string // optional: List method on the Client to use. Only required if we need to disambiguate between multiple options.
+	CustomListInput string // optional: string to set List input to (otherwise empty input will be used)
+
+	// Applies only to describe resources:
+	DescribeMethodName  string // optional: Describe method on the Client to use. Only required if we need to disambiguate between multiple options.
+	CustomDescribeInput string // optional: string to set List input to (otherwise empty input will be used)
+
+	// used for generating resolver and mock tests, but set automatically
+	parent   *Resource
+	children []*Resource
 }
 
 //go:embed templates/resolver_and_mock_test/*/*.go.tpl
@@ -146,6 +160,12 @@ func (r *Resource) Generate() error {
 	if r.Name != "" {
 		name = r.Name
 	}
+
+	// All table names must be plural
+	if !strings.HasSuffix(name, "s") {
+		return fmt.Errorf("invalid table name: %s. must be plural", name)
+	}
+
 	r.Table, err = codegen.NewTableFromStruct(
 		name,
 		r.Struct,
@@ -158,6 +178,10 @@ func (r *Resource) Generate() error {
 	r.Table.Resolver = "fetch" + strcase.ToCamel(r.Service) + strcase.ToCamel(r.SubService)
 	if r.Multiplex != "" {
 		r.Table.Multiplex = r.Multiplex
+		err = validateServiceMultiplex(r.Multiplex)
+		if err != nil {
+			return err
+		}
 	}
 	if r.PreResourceResolver != "" {
 		r.Table.PreResourceResolver = r.PreResourceResolver
@@ -225,12 +249,12 @@ func (r *Resource) generateResolver(dir string) error {
 	}).ParseFS(templatesFS,
 		fmt.Sprintf("templates/resolver_and_mock_test/%s/fetch.go.tpl", r.ResolverAndMockTestTemplate))
 	if err != nil {
-		return fmt.Errorf("failed to parse gcp templates: %w", err)
+		return fmt.Errorf("failed to parse templates: %w", err)
 	}
 
 	var buff bytes.Buffer
 	if err := tpl.Execute(&buff, r); err != nil {
-		return fmt.Errorf("failed to execute template: %w", err)
+		return fmt.Errorf("failed to execute resolver template for %s.%s: %w", r.Service, r.SubService, err)
 	}
 
 	filePath := path.Join(dir, r.SubService+"_fetch.go")
@@ -261,7 +285,7 @@ func (r *Resource) generateMockTest(dir string) error {
 
 	var buff bytes.Buffer
 	if err := tpl.Execute(&buff, r); err != nil {
-		return fmt.Errorf("failed to execute template: %w", err)
+		return fmt.Errorf("failed to execute mock template for %s.%s: %w", r.Service, r.SubService, err)
 	}
 
 	filePath := path.Join(dir, r.SubService+"_mock_test.go")
@@ -280,13 +304,105 @@ func (r *Resource) generateMockTest(dir string) error {
 	return nil
 }
 
+// SetParentChildRelationships calculates and sets the parent and children fields on resources.
+func SetParentChildRelationships(resources []*Resource) error {
+	m := map[string]*Resource{}
+	for _, r := range resources {
+		m[r.Service+"."+r.SubService] = r
+	}
+	csr := caser.New()
+	for _, r := range resources {
+		for _, ch := range r.Relations {
+			name := csr.ToSnake(strings.TrimSuffix(ch, "()"))
+			v, ok := m[r.Service+"."+name]
+			if !ok {
+				return errors.New("child not found for " + r.Service + "." + r.SubService + " : " + name)
+			}
+			r.children = append(r.children, v)
+			v.parent = r
+		}
+	}
+	return nil
+}
+
 // These methods are called from the template.
 // Because of this, we use a value receiver.
 // -------------------------------------------------------------------------------
 
-// Because usually the 'Struct' field contains a pointer, we need to dereference with '.Elem()'.
+// StructName returns the name of the resource's Struct field
 func (r Resource) StructName() string {
+	// because usually the 'Struct' field contains a pointer, we need to dereference with '.Elem()'.
 	return reflect.TypeOf(r.Struct).Elem().Name()
+}
+
+type ListTagsMethodResponse struct {
+	Method reflect.Method
+	Found  bool
+}
+
+// ListTagsMethod finds a ListTags method for the service, if any
+func (r Resource) ListTagsMethod() ListTagsMethodResponse {
+	if r.Client == nil || reflect.ValueOf(r.Client).IsNil() {
+		panic("Client needs to be set to generate resolvers and mocks")
+	}
+	m, err := discover.FindListTagsMethod(r.Client)
+	if err != nil {
+		return ListTagsMethodResponse{
+			Found: false,
+		}
+	}
+	return ListTagsMethodResponse{
+		Method: m.Method,
+		Found:  true,
+	}
+}
+
+// DescribeMethod finds a describe method for the resource
+func (r Resource) DescribeMethod() discover.DiscoveredMethod {
+	if r.Client == nil || (reflect.ValueOf(r.Client).Kind() == reflect.Ptr && reflect.ValueOf(r.Client).IsNil()) {
+		panic("Client needs to be set to generate resolvers and mocks")
+	}
+	if r.DescribeMethodName != "" {
+		m, err := discover.MethodByName(r.Client, r.Struct, r.DescribeMethodName)
+		if err != nil {
+			panic(err)
+		}
+		return m
+	}
+	m, err := discover.FindDescribeMethod(r.Client, r.Struct)
+	if err != nil {
+		panic(err)
+	}
+	return m
+}
+
+// ListMethod finds a list method for the resource
+func (r Resource) ListMethod() discover.DiscoveredMethod {
+	if r.Client == nil {
+		panic("Client needs to be set to generate resolvers and mocks")
+	}
+	if r.ListMethodName != "" {
+		m, err := discover.MethodByName(r.Client, r.Struct, r.ListMethodName)
+		if err != nil {
+			panic(err)
+		}
+		return m
+	}
+	m, err := discover.FindListMethod(r.Client, r.Struct)
+	if err != nil {
+		panic(err)
+	}
+	return m
+}
+
+// Parent returns the parent resource, if any
+func (r Resource) Parent() *Resource {
+	return r.parent
+}
+
+// Children returns the child resources, if any
+func (r Resource) Children() []*Resource {
+	return r.children
 }
 
 // CloudQueryServiceName is used for accessing 'client.Services().{{.CloudqueryServiceName}}' in templates
@@ -307,4 +423,28 @@ func CreateReplaceTransformer(replace map[string]string) func(field reflect.Stru
 		}
 		return name, nil
 	}
+}
+
+func validateServiceMultiplex(multiplexerCall string) error {
+	re := regexp.MustCompile(`\"(.*?)\"`)
+	// Find the value of the service parameter
+	submatchAll := re.FindStringSubmatch(multiplexerCall)
+	if len(submatchAll) != 2 {
+		return nil
+	}
+
+	t := client.ReadSupportedServiceRegions()
+	services := make(map[string]bool)
+
+	for _, partition := range t.Partitions {
+		for service := range partition.Services {
+			if _, ok := services[service]; !ok {
+				services[service] = true
+			}
+		}
+	}
+	if _, ok := services[submatchAll[1]]; !ok {
+		return fmt.Errorf("invalid partition: %s", submatchAll[1])
+	}
+	return nil
 }