+++ plugins/source/aws/codegen/services/clients.go
@@ -2,6 +2,7 @@ package services
 
 import (
 	"github.com/aws/aws-sdk-go-v2/service/accessanalyzer"
+	"github.com/aws/aws-sdk-go-v2/service/account"
 	"github.com/aws/aws-sdk-go-v2/service/acm"
 	"github.com/aws/aws-sdk-go-v2/service/apigateway"
 	"github.com/aws/aws-sdk-go-v2/service/apigatewayv2"
@@ -52,6 +53,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/inspector"
 	"github.com/aws/aws-sdk-go-v2/service/inspector2"
 	"github.com/aws/aws-sdk-go-v2/service/iot"
+	"github.com/aws/aws-sdk-go-v2/service/kafka"
 	"github.com/aws/aws-sdk-go-v2/service/kinesis"
 	"github.com/aws/aws-sdk-go-v2/service/kms"
 	"github.com/aws/aws-sdk-go-v2/service/lambda"
@@ -71,11 +73,13 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/aws/aws-sdk-go-v2/service/s3control"
 	"github.com/aws/aws-sdk-go-v2/service/sagemaker"
+	"github.com/aws/aws-sdk-go-v2/service/scheduler"
 	"github.com/aws/aws-sdk-go-v2/service/secretsmanager"
 	"github.com/aws/aws-sdk-go-v2/service/servicecatalog"
 	"github.com/aws/aws-sdk-go-v2/service/servicecatalogappregistry"
 	"github.com/aws/aws-sdk-go-v2/service/servicequotas"
 	"github.com/aws/aws-sdk-go-v2/service/sesv2"
+	"github.com/aws/aws-sdk-go-v2/service/sfn"
 	"github.com/aws/aws-sdk-go-v2/service/shield"
 	"github.com/aws/aws-sdk-go-v2/service/sns"
 	"github.com/aws/aws-sdk-go-v2/service/sqs"
@@ -92,6 +96,7 @@ import (
 
 var clients = []interface{}{
 	&accessanalyzer.Client{},
+	&account.Client{},
 	&acm.Client{},
 	&apigateway.Client{},
 	&apigatewayv2.Client{},
@@ -142,6 +147,7 @@ var clients = []interface{}{
 	&inspector.Client{},
 	&inspector2.Client{},
 	&iot.Client{},
+	&kafka.Client{},
 	&kinesis.Client{},
 	&kms.Client{},
 	&lambda.Client{},
@@ -161,11 +167,13 @@ var clients = []interface{}{
 	&s3.Client{},
 	&s3control.Client{},
 	&sagemaker.Client{},
+	&scheduler.Client{},
 	&secretsmanager.Client{},
 	&servicecatalog.Client{},
 	&servicecatalogappregistry.Client{},
 	&servicequotas.Client{},
 	&sesv2.Client{},
+	&sfn.Client{},
 	&shield.Client{},
 	&sns.Client{},
 	&sqs.Client{},