+++ plugins/source/aws/resources/services/glue/connections_fetch.go
@@ -2,8 +2,10 @@ package glue
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/aws/arn"
 	"github.com/aws/aws-sdk-go-v2/service/glue"
 	"github.com/aws/aws-sdk-go-v2/service/glue/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
@@ -31,8 +33,7 @@ func fetchGlueConnections(ctx context.Context, meta schema.ClientMeta, parent *s
 func resolveGlueConnectionArn(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
 	cl := meta.(*client.Client)
 	r := resource.Item.(types.Connection)
-	arn := aws.String(connectionARN(cl, &r))
-	return resource.Set(c.Name, arn)
+	return resource.Set(c.Name, connectionARN(cl, &r))
 }
 
 // ====================================================================================================================
@@ -40,5 +41,11 @@ func resolveGlueConnectionArn(ctx context.Context, meta schema.ClientMeta, resou
 // ====================================================================================================================
 
 func connectionARN(cl *client.Client, c *types.Connection) string {
-	return cl.ARN(client.GlueService, "connection", *c.Name)
+	return arn.ARN{
+		Partition: cl.Partition,
+		Service:   string(client.GlueService),
+		Region:    cl.Region,
+		AccountID: cl.AccountID,
+		Resource:  fmt.Sprintf("connection/%s", aws.ToString(c.Name)),
+	}.String()
 }