+++ plugins/source/terraform/CHANGELOG.md
@@ -5,6 +5,48 @@ All notable changes to this provider will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.24](https://github.com/cloudquery/cloudquery/compare/plugins-source-terraform-v1.2.23...plugins-source-terraform-v1.2.24) (2022-11-23)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/aws/aws-sdk-go-v2/config to v1.18.3 ([#4943](https://github.com/cloudquery/cloudquery/issues/4943)) ([e4aaf3f](https://github.com/cloudquery/cloudquery/commit/e4aaf3f0976a0836301b5de70a8e933c7abb5365))
+* **deps:** Update module github.com/aws/aws-sdk-go-v2/feature/s3/manager to v1.11.42 ([#4945](https://github.com/cloudquery/cloudquery/issues/4945)) ([e76c57d](https://github.com/cloudquery/cloudquery/commit/e76c57d49f44a1a370d1e2397b667be4df5c15c3))
+* **deps:** Update plugin-sdk for terraform to v1.7.0 ([#4916](https://github.com/cloudquery/cloudquery/issues/4916)) ([379a917](https://github.com/cloudquery/cloudquery/commit/379a917aef87e6e63ca6d59e95abd8d9a99dc8b5))
+* **deps:** Update plugin-sdk for terraform to v1.8.0 ([#4979](https://github.com/cloudquery/cloudquery/issues/4979)) ([e9faacc](https://github.com/cloudquery/cloudquery/commit/e9faacc16a4d37502ba9131fbe103537b3f6955e))
+
+## [1.2.23](https://github.com/cloudquery/cloudquery/compare/plugins-source-terraform-v1.2.22...plugins-source-terraform-v1.2.23) (2022-11-21)
+
+
+### Bug Fixes
+
+* **deps:** Update plugin-sdk for terraform to v1.6.0 ([#4854](https://github.com/cloudquery/cloudquery/issues/4854)) ([dca0c21](https://github.com/cloudquery/cloudquery/commit/dca0c21f8b68a0ded253db1448a220df808ecf5a))
+
+## [1.2.22](https://github.com/cloudquery/cloudquery/compare/plugins-source-terraform-v1.2.21...plugins-source-terraform-v1.2.22) (2022-11-18)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/aws/aws-sdk-go-v2/config to v1.18.1 ([#4736](https://github.com/cloudquery/cloudquery/issues/4736)) ([db70d2a](https://github.com/cloudquery/cloudquery/commit/db70d2a602fc4edfc74ed61fd7d28ada6da6a3af))
+* **deps:** Update module github.com/aws/aws-sdk-go-v2/config to v1.18.2 ([#4775](https://github.com/cloudquery/cloudquery/issues/4775)) ([136fb42](https://github.com/cloudquery/cloudquery/commit/136fb4213da150f8f9e4a68019fbe7fe94397370))
+* **deps:** Update module github.com/aws/aws-sdk-go-v2/credentials to v1.13.2 ([#4737](https://github.com/cloudquery/cloudquery/issues/4737)) ([2b2433d](https://github.com/cloudquery/cloudquery/commit/2b2433dd3e091147d0a8beed86159fa7b670e783))
+* **deps:** Update module github.com/aws/aws-sdk-go-v2/feature/s3/manager to v1.11.40 ([#4738](https://github.com/cloudquery/cloudquery/issues/4738)) ([69b01cc](https://github.com/cloudquery/cloudquery/commit/69b01cc65d9a62a4b734c956315a95aa1daafa86))
+* **deps:** Update module github.com/aws/aws-sdk-go-v2/feature/s3/manager to v1.11.41 ([#4776](https://github.com/cloudquery/cloudquery/issues/4776)) ([d0b2bae](https://github.com/cloudquery/cloudquery/commit/d0b2baeb33e047accc5ef0ac3b68f1605866fe4f))
+
+## [1.2.21](https://github.com/cloudquery/cloudquery/compare/plugins-source-terraform-v1.2.20...plugins-source-terraform-v1.2.21) (2022-11-16)
+
+
+### Bug Fixes
+
+* **deps:** Update plugin-sdk for terraform to v1.5.3 ([#4653](https://github.com/cloudquery/cloudquery/issues/4653)) ([b57598c](https://github.com/cloudquery/cloudquery/commit/b57598c904b65f97b9bc3592f534ad83c4ef5497))
+
+## [1.2.20](https://github.com/cloudquery/cloudquery/compare/plugins-source-terraform-v1.2.19...plugins-source-terraform-v1.2.20) (2022-11-15)
+
+
+### Bug Fixes
+
+* **terraform:** Update to new AWS SDK ([#4617](https://github.com/cloudquery/cloudquery/issues/4617)) ([1af16d1](https://github.com/cloudquery/cloudquery/commit/1af16d1a8aff80e458ee48593ba5ce7978938c35))
+
 ## [1.2.19](https://github.com/cloudquery/cloudquery/compare/plugins-source-terraform-v1.2.18...plugins-source-terraform-v1.2.19) (2022-11-15)